Second Project Using MVC
My website is a pharmacy with 7 branches. I have a list of medicines, each available at a specific branch. On the homepage, there is a welcome message for the user and two buttons that allow them to view branches and medicines. Users can perform several operations on the medicines, including adding a new medicine, deleting an existing one, viewing the full details of a specific medicine, editing it, or checking which branches have the medicine available. Additionally, users can search for a specific medicine; if it is found, its full details will be displayed, and if not, a message will indicate that it is unavailable. I first created the database, then used what was taught in LINQ and Entity Framework to manage the database, and finally used MVC to create the web pages.